requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
pillow==10.1.0
python-jose[cryptography]==3.3.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
prometheus-client==0.19.0
python-json-logger==2.0.7
imutils==0.5.4
scipy==1.11.4
psutil==5.9.6
scikit-learn==1.3.2
pytesseract==0.3.10
pdf2image==1.16.3
filterpy==1.4.5